requirements.txt
# Core server
fastapi==0.132.0
uvicorn==0.41.0
starlette==0.52.0

# HTTP clients
httpx==0.28.1
requests==2.32.5

# Ollama (LLM + embeddings)
ollama==0.6.1

# Vector search
faiss-cpu==1.13.2
numpy==2.4.2

# Retrieval / ranking
rank-bm25==0.2.2
beautifulsoup4==4.14.3
lxml==6.0.2

# NLP (entity detection)
spacy==3.8.11

# Config / validation
pydantic==2.12.5
PyYAML>=6.0
